add customRoutes endpoint (combination of user-defined and apps-defined routes)
