+++ CHANGELOG.md
@@ -9,9 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Public API endpoints for sitemap generation:
-  - `userRoutes`: Exposes user-defined routes from Rewriter
-  - `appsRoutes`: Exposes app-defined routes from store dependencies
+- Public API endpoint for sitemap generation:
+  - `customRoutes`: Exposes a combination of user-defined routes from Rewriter and app-defined routes from store dependencies.
 - Long cache control option for sitemap route endpoints
 
 ## [2.16.6] - 2025-01-30
