+++ docs/README.md
@@ -169,32 +169,26 @@ Beyond generating a complete sitemap, this app exposes specific API endpoints th
 
 #### Available Endpoints
 
-The following endpoints provide route information in JSON format:
+The following endpoint provides combined route information in JSON format:
 
-1. **User-defined Routes** 
-   - Endpoint: `/sitemap/user-routes`
-   - Description: Returns routes defined via CMS and internal pages (excluding product routes)
+1. **Custom Routes**
+   - Endpoint: `/_v/public/sitemap/custom-routes`
+   - Description: Returns a combination of routes defined via CMS and internal pages (excluding product routes) and routes defined by installed apps in their `build.json` files.
    - Example response:
    ```json
-   {
-     "routes": ["/about-us", "/contact", "/faq"],
-     "count": 3
-   }
+   [
+     {
+       "name": "apps-routes",
+       "routes": ["/store-locator/ny", "/store-locator/ca"]
+     },
+     {
+       "name": "user-routes",
+       "routes": ["/about-us", "/contact", "/faq"]
+     }
+   ]
    ```
 
-2. **App-defined Routes**
-   - Endpoint: `/sitemap/apps-routes`
-   - Description: Returns routes defined by installed apps in their `build.json` files
-   - Example response:
-   ```json
-   {
-     "routes": ["/store-locator/ny", "/store-locator/ca"],
-     "count": 2
-   }
-   ```
+#### Using the Endpoint
 
-#### Using the Endpoints
-
-These endpoints can be accessed via HTTP GET requests to:
-- `https://{workspace}--{account}.myvtex.com/sitemap/user-routes`
-- `https://{workspace}--{account}.myvtex.com/sitemap/apps-routes`
+This endpoint can be accessed via an HTTP GET request to:
+- `https://{workspace}--{account}.myvtex.com/_v/public/sitemap/custom-routes`
