+++ node/service.json
@@ -6,16 +6,6 @@
   "maxReplicas": 10,
   "workers": 4,
   "routes": {
-    "appsRoutes": {
-      "path": "/sitemap/apps-routes",
-      "public": true,
-      "smartcache": true
-    },
-    "userRoutes": {
-      "path": "/sitemap/user-routes",
-      "public": true,
-      "smartcache": true
-    },
     "sitemap": {
       "path": "/sitemap.xml",
       "public": true
@@ -31,6 +21,11 @@
     "robots": {
       "path": "/robots.txt",
       "public": true
+    },
+    "customRoutes": {
+      "path": "/_v/public/sitemap/custom-routes",
+      "public": true,
+      "smartcache": true
     }
   },
   "events": {
